+++ cspell.json
@@ -226,7 +226,8 @@
         "linea",
         "ifrandom",
         "WNATIVE",
-        "CRYPTOPUNKS"
+        "CRYPTOPUNKS",
+        "looksrare"
     ],
     "flagWords": [],
     "ignorePaths": [
