+++ src/app/(normal)/bookmarks/NFTBookmarkList.tsx
@@ -24,9 +24,9 @@ import { useCurrentProfileAll } from '@/hooks/useCurrentProfile.js';
 import { useIsLogin } from '@/hooks/useIsLogin.js';
 import { useNavigatorTitle } from '@/hooks/useNavigatorTitle.js';
 import { FireflySocialMediaProvider } from '@/providers/firefly/SocialMedia.js';
-import type { NftPreview } from '@/providers/types/Firefly.js';
+import type { SimpleHash } from '@/providers/types/SimpleHash.js';
 
-function getNFTItemContent(nft: NftPreview) {
+function getNFTItemContent(nft: SimpleHash.NFT) {
     const chainId = resolveSimpleHashChainId(nft.chain);
 
     const content = (
