+++ src/helpers/resolveNFTImageUrl.ts
@@ -1,8 +1,8 @@
 import { resolveImageURL } from '@masknet/web3-shared-evm';
 
-import type { NftPreview } from '@/providers/types/Firefly.js';
+import type { SimpleHash } from '@/providers/types/SimpleHash.js';
 
-export function resolveNFTImageUrl(nft: NftPreview) {
+export function resolveNFTImageUrl(nft: SimpleHash.NFT) {
     const original = nft.image_url || nft.previews.image_large_url;
 
     return nft.chain === 'solana'
